Moved dkg init command to standalone script in `/scripts/hooks` folder. Renamed `dkg` command to `ritual` command. Updated allowed values for --eth-staking-network and --coordinator-network to include `mainnet`/`polygon` respectively. Removed power of 2 restriction for dkg size.
